components moved out of pages
