+++ src/components/ViewSwitch.tsx
